Set WORKDIR and VOLUME, improve README
The `Dockerfile` didn't set up a WORKDIR, nor a VOLUME, so you were on your own when using in an ad-hoc way.  (CI/CD tools would tend to do okay, though.)  We now create/expose `/source` as the default mount point for the source to release.

Also fixed a few rough edges on the README.